+++ bindings/wasm/lib/index.ts
@@ -4,5 +4,7 @@
 import "./append_functions.js";
 export * from "./iota_identity_client.js";
 export * from "./jose";
+export * from "./jwk_storage";
+export * from "./key_id_storage";
 
 export * from "~identity_wasm";
